Only export the generic traced error version (#619)
This removes the boxed error type in favor of only exposing the generic error type and gates it behind a default-enabled feature flag.
