+++ tracing-error/Cargo.toml
@@ -34,8 +34,8 @@ keywords = [
 edition = "2018"
 
 [features]
-
-stack-error = []
+default = ["traced-error"]
+traced-error = []
 
 [dependencies]
 tracing-subscriber = { path = "../tracing-subscriber", version = "0.2.2", default-features = false, features = ["registry", "fmt"] }
